Support using as ESP IDF component by rename String and IPAddress class.
