+++ src/Packets/StringUtil.cpp
@@ -6,7 +6,7 @@ For a copy, see <https://opensource.org/licenses/MIT> or
 the LICENSE file.
 */
 
-#include "String.h"
+#include "StringUtil.h"
 
 namespace espMqttClientInternals {
