+++ src/Packets/StringUtil.h
