+++ src/Transport/IPAddressTest.cpp
@@ -8,7 +8,7 @@ the LICENSE file.
 
 #if defined(__linux__)
 
-#include "IPAddress.h"
+#include "IPAddressTest.h"
 
 IPAddress::IPAddress()
 : _address(0) {
